add pinned pyparsing version to reqs.txt
The regml.py script includes code that tries to import getTokensEndLoc
from the pyparsing module, but this seems to have been deprecated in the
most recent versions of that package.

The requirements.txt file in https://github.com/cfpb/regulations-parser/
has a pinned version to pyparsing==1.5.7, but the setup.py file in that
repo does not. So doing a clean install of just regulations-xml-parser
and doing `pip install -r requirements.txt` there will just download the
latest pyparsing, which breaks things and prevents regml.py from
running.

This would work if you already had the correct version of pyparsing
installed through some other means, but not if installing cleanly.

This change just adds the pinned version of pyparsing==1.5.7 to this
project's requirements.txt file so that it can be used in isolation
if that package isn't already installed.

All unit tests continue to pass via tox with this change.