modify settings to default to local XSD_FILE
This change modifies `settings.py` so that the `XSD_FILE` setting
defaults to a local directory as opposed to a copy of the schema hosted
on GitHub. This is necessary because lxml's URL-based loading doesn't
properly handle HTTPS URLs.

The new setting assumes a default of the schema repo checked out at the
same level as this repo, e.g. `../regulations-schema`. When the
validator schema is loaded, if the main schema file does not exist, a
descriptive error will be displayed to the user:

"Error occurred when reading schema file
../regulations-schema/src/eregs.xsd; did you forget to set
settings.XSD_FILE to a local or remote path containing the eregs schema?"

and the program will raise an exception and exit. This same error occurs
if the schema file is invalid somehow.

Some new unit tests have been added to validate this behavior. Because
the unit tests need the schema to test validation, the tests for the
validator now clone the GitHub repo locally before testing. This is
somewhat suboptimal but necessary for running in an automated way, e.g.
on Travis. This only happens once for the test class, not per test, and
things are cleaned up when the tests end.

Run `tox` to verify that test behavior works even without a preexisting
local copy of the schema.