add support for 1040 reg"letter"
This change adds support for regulation 1040 to the json-through
command. This command uses a lookup table to determine what should be
entered into the "regLetter" section of the RegML schema.

Regulation 1040 (arbitration reg) doesn't have a letter, but should be
referred to as "Regulation 1040", so this change just adds a hardcoded
1040 -> 1040 mapping.

To validate, run

```sh
./regml.py json-through 12 1040
```

to generate eRegs JSON from 1040 RegML.