add support for json-through start=-1
This change adds support for a `start=-1` parameter to the
`json-through` command which allows creation of JSON from RegML
automatically from the command line, with no user input necessary. This
allows for commands like

```
./regml.py json-through --start=-1 12 1026
```

which can be run via scripts to do long-running JSON generation.

Existing behavior to specify the start version is still supported.